Move graph component to component folder
